requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-nvidia-ai-endpoints>=0.1.0
openai>=1.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
